remove graphql package object
