+++ build.sbt
@@ -4,7 +4,7 @@ val clueVersion                = "0.20.3"
 val declineVersion             = "2.3.0"
 val disciplineMunitVersion     = "1.0.9"
 val flywayVersion              = "9.1.3"
-val grackleVersion             = "0.4.0"
+val grackleVersion             = "0.5.1"
 val http4sBlazeVersion         = "0.23.12"
 val http4sEmberVersion         = "0.23.14"
 val jwtVersion                 = "5.0.0"
