+++ modules/service/src/main/scala/lucuma/odb/graphql/binding/AngleBinding.scala
@@ -1,8 +1,7 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.odb.graphql
-package binding
+package lucuma.odb.graphql.binding
 
 import lucuma.core.math.Angle
 import lucuma.odb.graphql.util.Bindings._
