+++ modules/service/src/main/scala/lucuma/odb/graphql/binding/NonNegIntBinding.scala
@@ -1,0 +1,11 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql.binding
+
+import cats.syntax.all._
+import eu.timepit.refined.types.numeric.NonNegInt
+import lucuma.odb.graphql.util.Bindings._
+
+val NonNegIntBinding: Matcher[NonNegInt] =
+  IntBinding.emap(NonNegInt.from)
