+++ modules/service/src/main/scala/lucuma/odb/graphql/binding/PosBigDecimalBinding.scala
@@ -1,0 +1,13 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql.binding
+
+import cats.syntax.all._
+import eu.timepit.refined.types.numeric.PosBigDecimal
+import lucuma.odb.graphql.util.Bindings._
+
+val PosBigDecimalBinding: Matcher[PosBigDecimal] =
+  BigDecimalBinding.emap { s =>
+    PosBigDecimal.from(s).leftMap(m => s"Invalid PosBigDecimal: $s: $m")
+  }
