+++ modules/service/src/main/scala/lucuma/odb/graphql/binding/README.md
@@ -1,0 +1,3 @@
+
+This package contains bindings for leaf types (scalars and enums). Bindings for structured input
+types are in the `input` package.
