+++ modules/service/src/main/scala/lucuma/odb/graphql/input/AngleInput.scala
@@ -9,6 +9,7 @@ import cats.syntax.all._
 import edu.gemini.grackle.Result
 import lucuma.core.math.Angle
 import lucuma.core.math.HourAngle
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object AngleInput {
