+++ modules/service/src/main/scala/lucuma/odb/graphql/input/CatalogInfoInput.scala
@@ -7,6 +7,7 @@ package input
 import cats.syntax.all._
 import eu.timepit.refined.types.string.NonEmptyString
 import lucuma.core.enums.CatalogName
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 final case class CatalogInfoInput(
