+++ modules/service/src/main/scala/lucuma/odb/graphql/input/CreateObservationInput.scala
@@ -7,6 +7,7 @@ package input
 
 import cats.syntax.all._
 import lucuma.core.model.Program
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 final case class CreateObservationInput(
