+++ modules/service/src/main/scala/lucuma/odb/graphql/input/LinkUserInput.scala
@@ -7,6 +7,7 @@ package input
 
 import cats.syntax.all._
 import edu.gemini.grackle.Result
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 import lucuma.odb.service.ProgramService
 import lucuma.odb.service.ProgramService.LinkUserRequest
