+++ modules/service/src/main/scala/lucuma/odb/graphql/input/NonsiderealInput.scala
@@ -8,6 +8,7 @@ import cats.syntax.all._
 import edu.gemini.grackle.Result
 import lucuma.core.enums.EphemerisKeyType
 import lucuma.core.model.EphemerisKey
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object NonsiderealInput {
