+++ modules/service/src/main/scala/lucuma/odb/graphql/input/ProgramEditInput.scala
@@ -6,6 +6,7 @@ package lucuma.odb.graphql
 package input
 
 import lucuma.core.model.Program
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 case class ProgramEditInput(programId: Option[Program.Id])
