+++ modules/service/src/main/scala/lucuma/odb/graphql/input/ProgramPropertiesInput.scala
@@ -8,6 +8,7 @@ package input
 import cats.syntax.all._
 import eu.timepit.refined.types.string.NonEmptyString
 import lucuma.odb.data.Existence
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 case class ProgramPropertiesInput(
