+++ modules/service/src/main/scala/lucuma/odb/graphql/input/ProperMotionComponentInput.scala
@@ -12,6 +12,7 @@ import edu.gemini.grackle.Result
 import lucuma.core.math.ProperMotion
 import lucuma.core.math.VelocityAxis
 import lucuma.core.math.units._
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object ProperMotionComponentInput {
