+++ modules/service/src/main/scala/lucuma/odb/graphql/input/README.md
@@ -1,0 +1,3 @@
+
+This package contains bindings for structured input types. Bindings for leaf types (scalars and
+enums) are in the `bindings` package.
