+++ modules/service/src/main/scala/lucuma/odb/graphql/input/SetAllocationInput.scala
@@ -11,6 +11,7 @@ import lucuma.core.model.NonNegDuration
 import lucuma.core.model.Partner
 import lucuma.core.model.Program
 import lucuma.odb.data.Tag
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 case class SetAllocationInput(
