+++ modules/service/src/main/scala/lucuma/odb/graphql/input/UpdateProgramsInput.scala
@@ -8,6 +8,7 @@ package input
 import cats.syntax.all._
 import edu.gemini.grackle.Predicate
 import eu.timepit.refined.types.numeric.NonNegInt
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 case class UpdateProgramsInput(
