+++ modules/service/src/main/scala/lucuma/odb/graphql/input/WhereObservation.scala
@@ -11,6 +11,7 @@ import edu.gemini.grackle.Predicate
 import edu.gemini.grackle.Predicate._
 import lucuma.core.enums.ObsActiveStatus
 import lucuma.odb.data.ObsStatus
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object WhereObservation {
