+++ modules/service/src/main/scala/lucuma/odb/graphql/input/WhereOptionString.scala
@@ -13,6 +13,7 @@ import edu.gemini.grackle.Term
 import edu.gemini.grackle.sql.Like
 import eu.timepit.refined.cats._
 import eu.timepit.refined.types.string.NonEmptyString
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object WhereOptionString {
