+++ modules/service/src/main/scala/lucuma/odb/graphql/input/sourceprofile/FluxDensityContinuumInput.scala
@@ -12,6 +12,7 @@ import lucuma.core.math.BrightnessUnits._
 import lucuma.core.math.dimensional.Measure
 import lucuma.core.math.dimensional.Of
 import lucuma.core.math.dimensional.Units
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object FluxDensityContinuumInput {
