+++ modules/service/src/main/scala/lucuma/odb/graphql/input/sourceprofile/LineWidthInput.scala
@@ -8,6 +8,7 @@ package sourceprofile
 import coulomb.Quantity
 import eu.timepit.refined.types.numeric.PosBigDecimal
 import lucuma.core.math.units._
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object LineWidthInput {
