+++ modules/service/src/main/scala/lucuma/odb/graphql/input/sourceprofile/UnnormalizedSedInput.scala
@@ -13,6 +13,7 @@ import eu.timepit.refined.types.numeric
 import lucuma.core.enums.StellarLibrarySpectrum
 import lucuma.core.enums._
 import lucuma.core.model.UnnormalizedSED
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object UnnormalizedSedInput {
