+++ modules/service/src/main/scala/lucuma/odb/graphql/input/sourceprofile/WavelengthInput.scala
@@ -9,6 +9,7 @@ import cats.syntax.all._
 import edu.gemini.grackle.Result
 import lucuma.core.math.Wavelength
 import lucuma.core.optics.Format
+import lucuma.odb.graphql.binding._
 import lucuma.odb.graphql.util.Bindings._
 
 object WavelengthInput {
